double -> float

+++ src/types/type_system.rs
@@ -30,7 +30,7 @@ impl Display for Type {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
         match self {
             Type::Int => write!(f, "int"),
-            Type::Double => write!(f, "double"),
+            Type::Double => write!(f, "float"),
             Type::Bool => write!(f, "bool"),
             Type::Void => write!(f, "void"),
             Type::Error => write!(f, "?"),
